apiloader sets default orchestrator to DCOS if nil (#862)
* apiloader sets default orchestrator to DCOS if nil

This affects API versions v20160330, v20160930, and v20170131.

Also evolved the current tests in a minor way so we have a notion of a “default” spec. Currently the provided `defaults.json` specs in the version-specific `testdata` dirs have an empty (nil) orchestratorProfile. The invocations of api.setContainerServiceDefaultsv<api version>() in apiloader.go enable those default configurations to be processed.

* remove `defaults.json` from example test runs